Update EventHandlers.cs
增加了player为null的判断

+++ SystemNuke/EventHandlers.cs
@@ -19,6 +19,7 @@
                 Exiled.API.Features.Broadcast broadcast =new Exiled.API.Features.Broadcast($"SITE-02已被放弃，系统核弹将于{Warhead.RealDetonationTimer}秒后爆炸");
                 foreach ( Player p in Player.List)
                 {
+                    if (p == null) continue;
                     p.Broadcast( broadcast );
                 }
             }
@@ -32,11 +33,13 @@
                 Exiled.API.Features.Broadcast broadcast=new Exiled.API.Features.Broadcast($"为保证安全，基金会已发射核弹，将于{i}秒后到达", 1);
                 foreach (Player p in Player.List)
                 {
+                    if (p == null) continue;
                     p.Broadcast(broadcast);
                 }
             }
             foreach (Player p in Player.List)
             {
+                if (p == null) continue;
                 if (p.IsAlive)
                 {
                     p.Kill(DamageType.Warhead);
